test: add window isolation and host-helpers tests
Extract mapToolToMessage, mapComputerAction, formatToolContent to host-helpers.cjs for testability. Fixes dead code bug where _hint handling was unreachable in formatToolContent.

31 new tests covering window commands, tab windowId handling, and response formatting.
